create module in oap-server and move codes
