+++ .gitignore
@@ -19,7 +19,7 @@ OALLexer.tokens
 .externalToolBuilders
 oap-server/oal-grammar/**/gen/
 MQELexer.tokens
-oap-server/server-core/gen/
+oap-server/mqe-grammar/**/gen/
 PromQLLexer.tokens
 oap-server/server-query-plugin/promql-plugin/gen/
 LogQLLexer.tokens
