+++ oap-server/mqe-rt/src/main/java/org/apache/skywalking/mqe/rt/operation/aggregatelabels/SumAggregateLabelsFunc.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.mqe.rt.operation.aggregatelabels;
+
+public class SumAggregateLabelsFunc implements AggregateLabelsFunc {
+
+    private Double sum;
+
+    @Override
+    public void combine(final Double value) {
+        if (value == null) {
+            return;
+        }
+
+        if (sum == null) {
+            sum = value;
+        } else {
+            sum = sum + value;
+        }
+    }
+
+    @Override
+    public Double getResult() {
+        return sum;
+    }
+}
