--- oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/mqe/rt/type/ExpressionResultType.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.skywalking.oap.server.core.mqe.rt.type;
-
-public enum ExpressionResultType {
-
-    /**
-     * Can't resolve the type of the given expression.
-     */
-    UNKNOWN,
-    /**
-     * A single value
-     */
-    SINGLE_VALUE,
-    /**
-     * A collection of time-series values.
-     * The value could have labels or not.
-     */
-    TIME_SERIES_VALUES,
-    /**
-     * A collection of aggregated values through metric sort function.
-     */
-    SORTED_LIST,
-    /**
-     * A collection of sampled records.
-     * When the original metric type is sampled records.
-     */
-    RECORD_LIST
-}
